+++ packages/accounts-base/package.js
@@ -15,10 +15,6 @@ Package.onUse(api => {
   api.use('reactive-var', 'client');
   api.use('url', ['client', 'server']);
 
-  // use unordered to work around a circular dependency
-  // (service-configuration needs Accounts.connection)
-  api.use('service-configuration', ['client', 'server'], { unordered: true });
-
   // needed for getting the currently logged-in user and handling reconnects
   api.use('ddp', ['client', 'server']);
