+++ packages/accounts-oauth/oauth_server.js
@@ -1,3 +1,5 @@
+import { Meteor } from 'meteor/meteor';
+
 // Listen to calls to `login` with an oauth option set. This is where
 // users actually get logged in to meteor via oauth.
 Accounts.registerLoginHandler(options => {
@@ -54,4 +56,98 @@ Accounts.registerLoginHandler(options => {
     }
     return Accounts.updateOrCreateUserFromExternalService(result.serviceName, result.serviceData, result.options);
   }
+});
+
+///
+/// OAuth Encryption Support
+///
+
+const OAuthEncryption =
+  Package["oauth-encryption"] &&
+  Package["oauth-encryption"].OAuthEncryption;
+
+const usingOAuthEncryption = () => {
+  return OAuthEncryption && OAuthEncryption.keyIsLoaded();
+};
+
+// OAuth service data is temporarily stored in the pending credentials
+// collection during the oauth authentication process.  Sensitive data
+// such as access tokens are encrypted without the user id because
+// we don't know the user id yet.  We re-encrypt these fields with the
+// user id included when storing the service data permanently in
+// the users collection.
+//
+const pinEncryptedFieldsToUser = (serviceData, userId) => {
+  Object.keys(serviceData).forEach(key => {
+    let value = serviceData[key];
+    if (OAuthEncryption && OAuthEncryption.isSealed(value))
+      value = OAuthEncryption.seal(OAuthEncryption.open(value), userId);
+    serviceData[key] = value;
+  });
+};
+
+// Validate new user's email or Google/Facebook/GitHub account's email
+function defaultValidateNewUserHook(user) {
+  const domain = this._options.restrictCreationByEmailDomain;
+  if (!domain) {
+    return true;
+  }
+
+  let emailIsGood = false;
+  if (user.emails && user.emails.length > 0) {
+    emailIsGood = user.emails.reduce(
+      (prev, email) => prev || this._testEmailDomain(email.address), false
+    );
+  } else if (user.services && Object.values(user.services).length > 0) {
+    // Find any email of any service and check it
+    emailIsGood = Object.values(user.services).reduce(
+      (prev, service) => service.email && this._testEmailDomain(service.email),
+      false,
+    );
+  }
+
+  if (emailIsGood) {
+    return true;
+  }
+
+  if (typeof domain === 'string') {
+    throw new Meteor.Error(403, `@${domain} email required`);
+  } else {
+    throw new Meteor.Error(403, "Email doesn't match the criteria.");
+  }
+}
+
+// Encrypt unencrypted login service secrets when oauth-encryption is
+// added.
+//
+// XXX For the oauthSecretKey to be available here at startup, the
+// developer must call Accounts.config({oauthSecretKey: ...}) at load
+// time, instead of in a Meteor.startup block, because the startup
+// block in the app code will run after this accounts-base startup
+// block.  Perhaps we need a post-startup callback?
+
+Meteor.startup(() => {
+  if (! usingOAuthEncryption()) {
+    return;
+  }
+
+  Accounts._validateNewUserHooks = [
+    defaultValidateNewUserHook.bind(this)
+  ];
+
+  const { ServiceConfiguration } = Package['service-configuration'];
+
+  ServiceConfiguration.configurations.find({
+    $and: [{
+      secret: { $exists: true }
+    }, {
+      "secret.algorithm": { $exists: false }
+    }]
+  }).forEach(config => {
+    ServiceConfiguration.configurations.update(config._id, {
+      $set: {
+        secret: OAuthEncryption.seal(config.secret)
+      }
+    });
+  });
 });
