+++ packages/accounts-oauth/package.js
@@ -9,6 +9,11 @@ Package.onUse(api => {
   api.use(['accounts-base', 'ecmascript'], ['client', 'server']);
   // Export Accounts (etc) to packages using this one.
   api.imply('accounts-base', ['client', 'server']);
+
+  // use unordered to work around a circular dependency
+  // (service-configuration needs Accounts.connection)
+  api.use('service-configuration', ['client', 'server'], { unordered: true });
+
   api.use('oauth');
 
   api.addFiles('oauth_common.js');
